add all modules to app.module.ts
