+++ wee/webscraper/src/app/app.module.ts
@@ -2,11 +2,13 @@ import { Module } from '@nestjs/common';
 
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import {ScrapingModule } from '../industry-classification-app/industry.module';
+import { IndustryModule } from '../industry-classification-app/industry.module';
 import { RobotsModule } from '../robots-app/robots.module';
+import { StatusModule } from '../status-app/status.module';
+import { ImagesModule } from '../images-app/images.module';
 
 @Module({
-  imports: [RobotsModule,ScrapingModule],
+  imports: [RobotsModule, IndustryModule, StatusModule, ImagesModule],
   controllers: [AppController],
   providers: [AppService],
 })
