--- wee/webscraper/src/images-app/scraping.module.ts
@@ -1,11 +1,0 @@
-import { Module } from '@nestjs/common';
-import { ScrapingService } from './scraping.service';
-import { ScrapingController } from './scraping.controller';
-/**
- * Module for handling scraping-related functionality.
- */
-@Module({
-    providers: [ScrapingService],
-    controllers: [ScrapingController],
-})
-export class ScrapingModule {}
